Fix test setting flag that caused further tests to fail
